+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2.3.0 - 2023-09-12
+### Features
+* Added [`GetRemoteVisitorData`](https://developers.kameleoon.com/feature-management-and-experimentation/web-sdks/go-sdk/#GetRemoteVisitorData) method to fetch a visitor's remote data (with an optional capability to add the fetched data to the visitor).
+### Bug fixes
+* Stability and performance improvements
+
 ## 2.2.0 - 2023-08-20
 * Stability and performance improvements